AgentApp/pyproject.toml
[project]
name = "agentframework"
version = "0.1.0"
requires-python = ">=3.11,<3.13"
dependencies = [
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiolimiter==1.2.1",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anthropic==0.58.2",
    "anyio==4.10.0",
    "attrs==25.3.0",
    "authlib==1.6.3",
    "backoff==2.2.1",
    "basyx-python-sdk==1.2.1",
    "beartype==0.12.0",
    "beautifulsoup4==4.13.5",
    "blinker==1.9.0",
    "boto3==1.40.20",
    "botocore==1.40.20",
    "browser-use==0.1.40",
    "browsergym==0.13.3",
    "browsergym-assistantbench==0.13.3",
    "browsergym-core==0.13.3",
    "browsergym-experiments==0.13.3",
    "browsergym-miniwob==0.13.3",
    "browsergym-visualwebarena==0.13.3",
    "browsergym-webarena==0.13.3",
    "browsergym-workarena==0.4.4",
    "bubus==1.5.5",
    "cachetools==5.5.2",
    "cdp-use==1.4.0",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "cloudpickle==3.1.1",
    "colorama==0.4.6",
    "cryptography==45.0.6",
    "dataclasses-json==0.6.7",
    "datasets==3.4.1",
    "defusedxml==0.7.1",
    "dill==0.3.8",
    "distro==1.9.0",
    "docker==7.1.0",
    "duckduckgo-search==8.1.1",
    "english-words==2.0.2",
    "evaluate==0.4.5",
    "faiss-cpu==1.11.0",
    "faker==37.6.0",
    "farama-notifications==0.0.4",
    "fastapi==0.115.14",
    "filelock==3.19.1",
    "flask==3.1.2",
    "frozenlist==1.7.0",
    "fsspec==2024.12.0",
    "google-api-core==2.25.1",
    "google-api-python-client==2.179.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.2",
    "google-genai==1.29.0",
    "googleapis-common-protos==1.70.0",
    "googlesearch-python==1.3.0",
    "gradio==5.49.0",
    "greenlet==3.2.4",
    "groq==0.31.0",
    "gymnasium==1.1.1",
    "h11==0.16.0",
    "html2text==2024.2.26",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "huggingface-hub==0.34.4",
    "idna==3.10",
    "imageio==2.37.0",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "jmespath==1.0.1",
    "joblib==1.5.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.4.1",
    "langchain-anthropic==0.3.3",
    "langchain-core==0.3.75",
    "langchain-ollama==0.2.2",
    "langchain-openai==0.3.1",
    "langsmith==0.4.20",
    "lazy-loader==0.4",
    "libvisualwebarena==0.0.15",
    "libwebarena==0.0.4",
    "loguru==0.7.3",
    "lxml==5.4.0",
    "markdown-it-py==3.0.0",
    "markdown-pdf==1.5",
    "markdownify==0.14.1",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "matplotlib==3.10.7",
    "mcp==1.13.1",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "multidict==6.6.4",
    "multiprocess==0.70.16",
    "mypy-extensions==1.1.0",
    "networkx==3.5",
    "nltk==3.9.1",
    "numpy==2.3.2",
    "oauthlib==3.3.1",
    "ollama==0.5.3",
    "openai==1.99.2",
    "orjson==3.11.3",
    "packaging==25.0",
    "pandas==2.3.2",
    "pdf2image==1.17.0",
    "pillow==11.1.0",
    "playwright==1.55.0",
    "portalocker==2.10.1",
    "posthog==6.7.0",
    "prettytable==3.16.0",
    "primp==0.15.0",
    "propcache==0.3.2",
    "proto-plus==1.26.1",
    "protobuf==6.32.0",
    "psutil==7.0.0",
    "pyarrow==21.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pydub==0.25.1",
    "pyecma376-2==1.0.1",
    "pyee==13.0.0",
    "pygments==2.19.2",
    "pymupdf==1.25.3",
    "pyparsing==3.2.3",
    "pypdf==6.0.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pywin32==311",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "regex==2025.7.34",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==14.1.0",
    "rpds-py==0.27.1",
    "rsa==4.9.1",
    "s3transfer==0.13.1",
    "sacrebleu==2.5.1",
    "safetensors==0.6.2",
    "scikit-image==0.25.2",
    "scikit-learn==1.7.0",
    "scipy==1.16.1",
    "screeninfo==0.8.1",
    "seaborn==0.13.2",
    "setuptools==75.8.2",
    "shellingham==1.5.4",
    "six==1.17.0",
    "skops==0.13.0",
    "sniffio==1.3.1",
    "soupsieve==2.8",
    "sse-starlette==3.0.2",
    "starlette==0.46.2",
    "sympy==1.14.0",
    "tabulate==0.9.0",
    "tenacity==9.0.0",
    "text-generation==0.7.0",
    "tifffile==2025.8.28",
    "tiktoken==0.9.0",
    "tokenizers==0.21.4",
    "tomli==2.2.1",
    "torch==2.8.0",
    "tqdm==4.67.1",
    "transformers==4.55.4",
    "trimesh==4.7.4",
    "typer==0.16.1",
    "types-requests==2.32.4.20250809",
    "types-tqdm==4.67.0.20250809",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "unidiff==0.7.5",
    "uritemplate==4.2.0",
    "urllib3==2.3.0",
    "uuid7==0.1.0",
    "uvicorn==0.34.3",
    "weblinx==0.3.2",
    "weblinx-browsergym==0.0.3",
    "websockets==15.0.1",
    "werkzeug==3.1.3",
    "win32-setctime==1.2.0",
    "xgboost==3.1.0",
    "xxhash==3.5.0",
    "yarl==1.20.1",
    "zstandard==0.24.0",
]

[project.optional-dependencies]
dev = [
    "ruff==0.14.1",
]
